Feature rocketsound database

+++ modules/telestion-core/src/main/java/de/jvpichowski/rocketsound/messages/base/Altitude.java
@@ -3,10 +3,11 @@ package de.jvpichowski.rocketsound.messages.base;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.telestion.api.message.JsonMessage;
 
-public record Altitude (@JsonProperty float temp) implements JsonMessage {
+// TODO: float not yet supported by vertx -> changed floats to double (Jan)
+public record Altitude(@JsonProperty double temp) implements JsonMessage {
 
 	@SuppressWarnings("unused")
 	public Altitude(){
-		this(0.0f);
+		this(0.0);
 	}
 }
