+++ modules/telestion-core/src/main/java/de/jvpichowski/rocketsound/messages/base/GpsData.java
@@ -3,12 +3,15 @@ package de.jvpichowski.rocketsound.messages.base;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.telestion.api.message.JsonMessage;
 
-public record GpsData(@JsonProperty int satCount, @JsonProperty int fix,
-					  @JsonProperty float north, @JsonProperty float east,
-					  @JsonProperty long time) implements JsonMessage {
-
+// TODO: float not yet supported by vertx -> changed floats to double (Jan)
+public record GpsData(
+		@JsonProperty int satCount,
+		@JsonProperty int fix,
+		@JsonProperty double north,
+		@JsonProperty double east,
+		@JsonProperty long time) implements JsonMessage {
 	@SuppressWarnings("unused")
 	public GpsData(){
-		this(0, 0, 0.0f, 0.0f, 0);
+		this(0, 0, 0.0, 0.0, 0);
 	}
 }
